Move Moshi integration behind private module
Remove the public-looking mobius.integrations.moshi package with no compatibility shim. Keep PersonaPlex support solely in the private integration module used by the standard mobius.build and CLI routes.


Copilot-Session: 43bb31ac-c136-44bc-8cdf-70cbcedd4a42
